implemented on typescript
