+++ ethers-simple-storage/encryptKey.ts
@@ -1,0 +1,19 @@
+import { ethers } from "ethers";
+import * as fs from "fs";
+import "dotenv/config";
+
+async function main() {
+  const wallet = new ethers.Wallet(process.env.PRIVATE_KEY!);
+  const encryptedJsonKey = await wallet.encrypt(
+    process.env.PRIVATE_KEY_PASSWORD!
+  );
+  console.log(encryptedJsonKey);
+  fs.writeFileSync("./.encryptedJsonKey.json", encryptedJsonKey);
+}
+
+main()
+  .then(() => process.exit(0))
+  .catch((error) => {
+    console.error(error);
+    process.exit(1);
+  });
