+++ ethers-simple-storage/package.json
@@ -1,12 +1,17 @@
 {
   "dependencies": {
     "dotenv": "^16.1.4",
-    "ethers": "^6.5.1",
+    "ethers": "5.7.2",
     "prettier": "^2.8.8",
     "prettier-plugin-solidity": "^1.1.3",
-    "solc": "0.8.7-fixed"
+    "solc": "0.8.7-fixed",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.3"
   },
   "scripts": {
     "compile": "yarn solcjs --bin --abi --include-path node_modules/ --base-path . -o . SimpleStorage.sol "
+  },
+  "devDependencies": {
+    "@types/node": "^20.3.1"
   }
 }
